move minimize def
